Move reset top left funct to gfx utils
