+++ include/layout.h
@@ -11,5 +11,8 @@
 static const Rect TOP_LEFT_PANEL_ANIM_RECT = {0, 0, 8, 4};
 static const Rect POP_MENU_ANIM_RECT = {9, 7, 24, 31};
 static const BG_POINT CUR_BLIND_TOKEN_POS = {8, 18};
+static const BG_POINT TOP_LEFT_PANEL_POINT  = {0,       0, };
+static const Rect TOP_LEFT_PANEL_BOTTOM_ROW_RESET_RECT = {0, 28, 8,     28 };
+static const Rect TOP_LEFT_ITEM_SRC_RECT    = {0,       20,     8,      25 };
 
 #endif // LAYOUT_H
